debug timer piles

+++ largerpile.cpp
@@ -5,7 +5,7 @@
 LargerPile::LargerPile(Player* p,QVector2D pos,QVector2D text):Pile(p,pos,7,30,90,1,text,1,QVector3D(0,3,0)){
 }
 
-LargerPile::LargerPile(Player* p,QVector2D pos,float r,float c,float l,int d,QVector2D text):Pile(p,pos,r,c,l,d,text,1,QVector3D(0,3,0)){
+LargerPile::LargerPile(Player* p,QVector2D pos,float r,float c,int l,int d,QVector2D text):Pile(p,pos,r,c,l,d,text,1,QVector3D(0,3,0)){
 
 }
