+++ largerpile.h
@@ -10,7 +10,7 @@ class LargerPile : public Pile{
 public :
 
    LargerPile(Player* p,QVector2D pos,QVector2D text);
-   LargerPile(Player* p,QVector2D pos,float r,float c, float l,int d,QVector2D text);
+   LargerPile(Player* p,QVector2D pos,float r,float c, int l,int d,QVector2D text);
    void Affect(Movable* m) override;
 };
