+++ mainpile.cpp
@@ -6,7 +6,7 @@ MainPile::MainPile(Player* p,QVector2D pos,QVector2D text):Pile(p,pos,4,50,60,0,
 
 }
 
-MainPile::MainPile(Player* p,QVector2D pos,float r, float c, float l, int d,QVector2D text):Pile(p,pos,r,c,l,d,text,0,QVector3D(3,3,0)){
+MainPile::MainPile(Player* p,QVector2D pos,float r, float c, int l, int d,QVector2D text):Pile(p,pos,r,c,l,d,text,0,QVector3D(3,3,0)){
 
 }
