+++ mainpile.h
@@ -7,7 +7,7 @@ class MainPile : public Pile{
 
 public :
    MainPile(Player* p,QVector2D pos,QVector2D text);
-   MainPile(Player* p,QVector2D pos,float r,float c, float l,int d,QVector2D text);
+   MainPile(Player* p,QVector2D pos,float r,float c, int l,int d,QVector2D text);
    void Affect(Movable* m) override;
 };
