+++ rangedpile.cpp
@@ -6,7 +6,7 @@ RangedPile::RangedPile(Player* p,QVector2D pos,QVector2D text):Pile(p,pos,6.25,5
 
 }
 
-RangedPile::RangedPile(Player* p,QVector2D pos,float r, float c, float l, int d,QVector2D text):Pile(p,pos,r,c,l,d,text,0,QVector3D(3,0,0)){
+RangedPile::RangedPile(Player* p,QVector2D pos,float r, float c, int l, int d,QVector2D text):Pile(p,pos,r,c,l,d,text,0,QVector3D(3,0,0)){
 
 }
